tuning

+++ setup.py
@@ -4,60 +4,33 @@ import os
 import shutil
 # Dependencies are automatically detected, but it might need
 # fine tuning.
-excludes = [
-    'tkinter',
-    'zmq',
-    'sqlite3',
-    'gevent',
-    'greenlet',
-    'coverage',
-    'lzma',
-    'cffi',
-    'asyncio',
-    'bz2',
-    'jinja2',
-]
 zip_exclude_packages = [
     'certifi',
     'pydivert',
 ]
-buildOptions = dict(packages=[], excludes=excludes, includes=['idna.idnadata', 'codecs', 'pydivert'], replace_paths=[("*", "")],
-                    optimize=2, zip_include_packages="*", zip_exclude_packages=zip_exclude_packages, silent=True)
-base = 'Console'
-
+buildOptions = dict(packages=[], includes=['idna.idnadata', 'codecs', 'pydivert'],
+                    replace_paths=[("*", "")], optimize=2, zip_include_packages="*",
+                    zip_exclude_packages=zip_exclude_packages, silent=True)
 executables = [
-    Executable('app.py', base=base, targetName='Guardian.exe', icon='logo.ico')
+    Executable('app.py', targetName='Guardian.exe', icon='logo.ico')
 ]
 
 version = '3.0.2'
 
-try:
+if os.path.exists('build/exe.win-amd64-3.6'):
+    shutil.rmtree('build/exe.win-amd64-3.6')
+
+if not os.path.exists('build/exe'):
+    os.makedirs('build/exe')
+
+if os.path.isfile('build/exe/guardian-{}.zip'.format(version)):
     os.remove('build/exe/guardian-{}.zip'.format(version))
-except:
-    pass
-try:
-    os.remove('build/exe.win-amd64-3.6/Guardian.exe')
-except:
-    pass
-try:
-    shutil.rmtree('build/exe.win-amd64-3.6/lib')
-except:
-    pass
-try:
-    os.remove('build/exe.win-amd64-3.6/python36.dll')
-except:
-    pass
-try:
-    shutil.copyfile('LICENSE', 'build/exe.win-amd64-3.6/LICENSE')
-    shutil.copyfile('SOURCE', 'build/exe.win-amd64-3.6/SOURCE')
-except:
-    pass
+
 setup(name='Guardian',
       version=version,
       description='Firewall',
       options=dict(build_exe=buildOptions),
       executables=executables)
-
 
 def zip_folder(folder_path, output_path):
     """Zip the contents of an entire folder (with that folder included
@@ -84,7 +57,8 @@ def zip_folder(folder_path, output_path):
     zip_file.close()
 
 try:
-    os.remove('build/exe.win-amd64-3.6/lib/python36.dll')
+    shutil.copyfile('LICENSE', 'build/exe.win-amd64-3.6/LICENSE')
+    shutil.copyfile('SOURCE', 'build/exe.win-amd64-3.6/SOURCE')
 except:
     pass
